RDK-34105: Script changes
Reason for change: Adding more info about the crashed process
Test Procedure: Kill any process and check the respective log files
                are added to the minidump tar ball
Risks: Low

(cherry picked from commit b9f79350698002164be05d621f72a177dc839c17)

+++ uploadDumps.sh
@@ -21,6 +21,8 @@
 
 #This file is from crashupload repository
 #Uploads coredumps to an ftp server if there are any
+LOGMAPPER_FILE="/etc/breakpad-logmapper.conf"
+LOG_FILES="/tmp/minidump_log_files.txt"
 
 if [ -f /lib/rdk/t2Shared_api.sh ]; then
     source /lib/rdk/t2Shared_api.sh
@@ -1184,11 +1186,16 @@ logFileCopy()
        line_count=500
     fi
 
-    if [ ! -z "$STBLOG_FILE" -a -f "$STBLOG_FILE" ]; then
-        tail -n ${line_count} $STBLOG_FILE > $stbLogFile
-    fi
-    if [ ! -z "$OCAPLOG_FILE" -a -f "$OCAPLOG_FILE" ]; then
-        tail -n ${line_count} $OCAPLOG_FILE > $ocapLogFile
+    if [ "$DUMP_FLAG" != "0" ]; then
+        if [ ! -z "$STBLOG_FILE" -a -f "$STBLOG_FILE" ]; then
+            tail -n ${line_count} $STBLOG_FILE > $stbLogFile
+        fi
+        if [ ! -z "$OCAPLOG_FILE" -a -f "$OCAPLOG_FILE" ]; then
+            tail -n ${line_count} $OCAPLOG_FILE > $ocapLogFile
+        fi
+        if [ ! -z "$WPEFRAMEWORK_LOG" -a -f "$WPEFRAMEWORK_LOG" ]; then
+            tail -n ${line_count} $WPEFRAMEWORK_LOG > $wpeLogFile
+        fi
     fi
     if [ ! -z "$MESSAGE_TXT" -a -f "$MESSAGE_TXT" ]; then
         tail -n ${line_count} $MESSAGE_TXT > $messagesTxtFile
@@ -1201,9 +1208,6 @@ logFileCopy()
     fi
     if [ ! -z "$CEF_LOG" -a -f "$CEF_LOG" ]; then
         tail -n 500 $CEF_LOG > $cefLogFile
-    fi
-    if [ ! -z "$WPEFRAMEWORK_LOG" -a -f "$WPEFRAMEWORK_LOG" ]; then
-        tail -n ${line_count} $WPEFRAMEWORK_LOG > $wpeLogFile
     fi
 }
 
@@ -1233,6 +1237,20 @@ shouldProcessFile()
     fi
 }
 
+add_crashed_log_file()
+{
+    file="$1"
+    pname=`echo $file | awk -F_ '{print $1}'`
+    pname=${pname#"./"} #Remove ./ from the dump name
+    logMessage "Process crashed = $pname"
+    get_logs=`awk -v proc="$pname" -F= '$1 ~ proc {print $2}' $LOGMAPPER_FILE`
+    logMessage "Crashed process log file: $get_logs"
+    for i in $(echo $get_logs | sed -n 1'p' | tr ',' '\n'); do
+        echo "$LOG_PATH/$i" >> $LOG_FILES
+    done
+    logMessage "Logs: `cat $LOG_FILES`"
+}
+
 processDumps()
 {
     # wait for app buffers are flushed
@@ -1247,6 +1265,9 @@ processDumps()
         elif [ "$f1" != "$f" ]; then
             mv "$f" "$f1"
             f="$f1"
+        fi
+        if [ "$DUMP_FLAG" == "0" ]; then
+            add_crashed_log_file "$f"
         fi
         if [ -f "$f" ]; then
             #last modification date of a core dump, to ease refusing of already uploaded core dumps on a server side
@@ -1323,7 +1344,7 @@ processDumps()
                 fi
             else
                 if [ "$DEVICE_TYPE" = "hybrid" ] || [ "$DEVICE_TYPE" = "mediaclient" ]; then
-                    files="$tgzFile $dumpName $VERSION_FILE $stbLogFile $ocapLogFile $messagesTxtFile $appStatusLogFile $appLogFile $cefLogFile $wpeLogFile $CORE_LOG $crashedUrlFile"
+                    files="$tgzFile $dumpName $VERSION_FILE $messagesTxtFile $appStatusLogFile $appLogFile $cefLogFile $CORE_LOG $crashedUrlFile"
                     if [ "$BUILD_TYPE" != "prod" ]; then
                         test -f $LOG_PATH/receiver.log && files="$files $LOG_PATH/receiver.log*"
                         test -f $LOG_PATH/thread.log && files="$files $LOG_PATH/thread.log"
@@ -1331,6 +1352,18 @@ processDumps()
                         test -f $LOG_PATH/receiver.log && files="$files $LOG_PATH/receiver.log"
                         test -f $LOG_PATH/receiver.log.1 && files="$files $LOG_PATH/receiver.log.1"
                     fi
+                    while read line
+                    do
+                        if [ ! -z "$line" -a -f "$line" ]; then
+                            logModTS=`getLastModifiedTimeOfFile $line`
+                            checkParameter logModTS
+                            process_log=`setLogFile $sha1 $MAC $logModTS $boxType $modNum $line`
+                            tail -n ${line_count} $line > $process_log
+                            logMessage "Adding File: $process_log to minidump tarball"
+                            files="$files $process_log"
+                        fi
+                    done < $LOG_FILES
+                    rm -rf $LOG_FILES
                     nice -n 19 tar -zcvf $files 2>&1 | logStdout
                     if [ $? -eq 0 ]; then
                         logMessage "Success Compressing the files $files"
@@ -1352,13 +1385,19 @@ processDumps()
 
             rm $dumpName
             if [ "$DEVICE_TYPE" = "hybrid" ] || [ "$DEVICE_TYPE" = "mediaclient" ]; then
-                if [ ! -z "$STBLOG_FILE" -a -f "$STBLOG_FILE" ]; then
-                    logMessage "Removing $stbLogFile"
-                    rm $stbLogFile
-                fi
-                if [ ! -z "$OCAPLOG_FILE" -a -f "$OCAPLOG_FILE" ]; then
-                    logMessage "Removing $ocapLogFile"
-                    rm $ocapLogFile
+                if [ "$DUMP_FLAG" != "0" ]; then
+                    if [ ! -z "$STBLOG_FILE" -a -f "$STBLOG_FILE" ]; then
+                        logMessage "Removing $stbLogFile"
+                        rm $stbLogFile
+                    fi
+                    if [ ! -z "$OCAPLOG_FILE" -a -f "$OCAPLOG_FILE" ]; then
+                        logMessage "Removing $ocapLogFile"
+                        rm $ocapLogFile
+                    fi
+                    if [ ! -z "$WPEFRAMEWORK_LOG" -a -f "$WPEFRAMEWORK_LOG" ]; then
+                        logMessage "Removing $wpeLogFile"
+                        rm $wpeLogFile
+                    fi
                 fi
                 if [ ! -z "$MESSAGE_TXT" -a -f "$MESSAGE_TXT" ]; then
                     logMessage "Removing $messagesTxtFile"
@@ -1380,9 +1419,9 @@ processDumps()
                     logMessage "Removing $cefLogFile"
                     rm $cefLogFile
                 fi
-                if [ ! -z "$WPEFRAMEWORK_LOG" -a -f "$WPEFRAMEWORK_LOG" ]; then
-                    logMessage "Removing $wpeLogFile"
-                    rm $wpeLogFile
+                if [ "$DUMP_FLAG" == "0" ]; then
+                    process_logs=`find $WORKING_DIR \( -iname "*.log" -o -iname "*.txt" \) -type f -print -exec rm -f {} \;`
+                    logMessage "Removing ${process_logs}"
                 fi
             fi
         fi
